requirements.txt
# NOTE: This project is used on macOS arm64 with Python 3.13.
# torch==2.2.2 has no wheels for Python 3.13; use a compatible newer build.
torch==2.6.0
# transformers 4.44.x pulls tokenizers 0.19.x which has no cp313 wheels (would compile from source).
# 4.45+ supports tokenizers 0.20.x with cp313 wheels.
transformers==4.45.0
tokenizers==0.20.3
sentence-transformers==3.0.1
huggingface-hub==0.31.4
# faiss-cpu==1.8.0.post1 does not publish Python 3.13 wheels; use a newer build.
faiss-cpu==1.13.2
fastapi==0.116.1
uvicorn==0.35.0
sqlalchemy==2.0.43
psycopg[binary]==3.2.10
PyJWT==2.10.1
pgvector==0.3.6
pypdf==5.1.0
requests>=2.31.0
python-dotenv==1.0.1
